-Get Total Houses By Category Id Impl (#197)

+++ src/main/kotlin/com/realtor/plugins/dao/house/HousesDao.kt
@@ -54,7 +54,7 @@ interface HousesDao {
     suspend fun getHousesById(id: Long): Houses?
     suspend fun getHouseByCategoryId(id: Long): Houses?
     suspend fun getHousesListByCategoryId(id: Long): List<Houses>?
-    suspend fun getTotalHousesByCategoryId(categoryId: Long): Int?
+    suspend fun getTotalHousesByCategoryId(categoryId: Long): Long?
     suspend fun deleteHouseByCategoryId(id: Long): Int?
     suspend fun deleteHouseById(id: Long): Int?
     suspend fun updateHouseById(
